Move utilities from dask to rabit.
